requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9

faiss-cpu==1.8.0.post1
beautifulsoup4==4.12.3
pypdf==4.3.1
sentence-transformers==3.0.1
openai==0.28.1
pydantic==2.9.2
pydantic-settings==2.5.2
